Converted all double to big decimals

+++ dps-web-service/src/main/java/com/dps/web/service/controller/DashboardController.java
@@ -49,7 +49,7 @@ public class DashboardController
 		
 		DashboardDTO dashboardDto = new DashboardDTO();
 		dashboardDto.setCustomerCount(customerCount);
-		dashboardDto.setExchangeRate(config.getExchangeRate().doubleValue());
+		dashboardDto.setExchangeRate(config.getExchangeRate());
 		dashboardDto.setProductsCount(productCount);
 		dashboardDto.setSupplierCount(supplierCount);
